Emphasize importance of deploying after merge
Past couple of days has seen a couple of PRs that were merged, but not immediately deployed.

This leaves any potential issues as nice little surprises for the next deployer to discover and fix! :sweat_smile:

Even after getting a :+1:, if aren't going to be around when the code gets deployed, better to wait to merge till a time where you can follow up on any problems that might occur.