Update README.md
I have recently been fixing a bug on the counter which was incrementing the recipe photo_comment count using `dismissed` new comments, and then re-incrementing when the comment gets `active`. The issue ended up with the photo comment being counted twice in the recipe rather than once. 
The test was there, but since it was using the `expect` `change` Rspec syntax with a `by(1)`, we haven't noticed that the counter which should have been going from 0 to 1, was actually going from 1 to 2. 
Hence, I believe is valuable to add this to our testing style guide