Disabled SuggestExtensions
Why:

SuggestedExtensions displays a message like

`
The following RuboCop extension libraries are installed but not loaded in config:
`

every time someone runs Rubocop.

If you run Rubocop on multiple files this message will be printed each time
Rubocop runs, thus making the output of the analysis hard to follow.

I think it is better to let the output of Rubocop remain focused on
reporting offences.